Update malicious spdz mac check in tests
