+++ core/src/test/java/dk/alexandra/fresco/lib/generic/TestCommitmentComputation.java
@@ -5,20 +5,15 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import dk.alexandra.fresco.commitment.HashBasedCommitment;
 import dk.alexandra.fresco.commitment.HashBasedCommitmentSerializer;
 import dk.alexandra.fresco.framework.Application;
-import dk.alexandra.fresco.framework.DRes;
 import dk.alexandra.fresco.framework.MaliciousException;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
 import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
 import dk.alexandra.fresco.framework.builder.numeric.ProtocolBuilderNumeric;
-import dk.alexandra.fresco.framework.network.serializers.ByteSerializer;
 import dk.alexandra.fresco.framework.sce.evaluator.EvaluationStrategy;
 import dk.alexandra.fresco.framework.sce.resources.ResourcePool;
 import dk.alexandra.fresco.framework.util.AesCtrDrbg;
-import dk.alexandra.fresco.framework.util.Drbg;
-import dk.alexandra.fresco.framework.util.Pair;
 import dk.alexandra.fresco.suite.dummy.arithmetic.AbstractDummyArithmeticTest;
 import java.util.ArrayList;
 import java.util.List;
@@ -115,7 +110,7 @@ public class TestCommitmentComputation extends AbstractDummyArithmeticTest {
               root -> {
                 final int myId = root.getBasicNumericContext().getMyId();
                 if (myId == cheatingPartyId) {
-                  return new MaliciousTwoPartyCommitmentComputation(
+                  return new MaliciousCommitmentComputation(
                       commitmentSerializer,
                       inputs.get(myId - 1), noParties,
                       localDrbg)
@@ -143,50 +138,6 @@ public class TestCommitmentComputation extends AbstractDummyArithmeticTest {
           }
         }
       };
-    }
-  }
-
-  static class MaliciousTwoPartyCommitmentComputation extends CommitmentComputation {
-
-    private final ByteSerializer<HashBasedCommitment> commitmentSerializer;
-    private final byte[] value;
-    private final int noOfParties;
-    private final Drbg localDrbg;
-
-    MaliciousTwoPartyCommitmentComputation(
-        ByteSerializer<HashBasedCommitment> commitmentSerializer,
-        byte[] value, int noOfParties, Drbg localDrbg) {
-      super(commitmentSerializer, value, noOfParties, localDrbg);
-      this.commitmentSerializer = commitmentSerializer;
-      this.value = value;
-      this.noOfParties = noOfParties;
-      this.localDrbg = localDrbg;
-    }
-
-    @Override
-    public DRes<List<byte[]>> buildComputation(ProtocolBuilderNumeric builder) {
-      HashBasedCommitment ownCommitment = new HashBasedCommitment();
-      byte[] ownOpening = ownCommitment.commit(localDrbg, value);
-      return builder.seq(
-          seq -> {
-            if (noOfParties > 2) {
-              return new BroadcastComputation<ProtocolBuilderNumeric>(
-                  commitmentSerializer.serialize(ownCommitment))
-                  .buildComputation(seq);
-            } else {
-              return seq.append(new InsecureBroadcastProtocol<>(
-                  commitmentSerializer.serialize(ownCommitment)));
-            }
-          })
-          .seq((seq, rawCommitments) -> {
-            // tamper with opening
-            ownOpening[1] = (byte) (ownOpening[1] ^ 1);
-            DRes<List<byte[]>> res = seq.append(new InsecureBroadcastProtocol<>(ownOpening));
-            final Pair<DRes<List<byte[]>>, List<byte[]>> dResListPair = new Pair<>(res,
-                rawCommitments);
-            return () -> dResListPair;
-          })
-          .seq((seq, pair) -> null);
     }
   }
